+++ src/@types/raw.macro.d.ts
@@ -1,4 +1,3 @@
-
-declare module "raw.macro" {
+declare module 'raw.macro' {
   export default function (pathname: string): string;
 }