+++ src/db/drawlet-cache.ts
@@ -1,4 +1,4 @@
-import DrawingModel from './file-format/DrawingModel';
+import DrawingModel from '../lascaux/data-model/DrawingModel';
 import { getOrAdd, PromiseOrValue } from 'promise-or-value';
 
 const map = new Map<string, DrawingModel | Error>();