+++ src/lascaux/util/typed-arrays.ts
@@ -1,0 +1,13 @@
+export type TypedArray =
+  | Uint8Array
+  | Uint8ClampedArray
+  | Uint16Array
+  | Uint32Array
+  | Int8Array
+  | Int16Array
+  | Int32Array
+  | Float32Array
+  | Float64Array;
+export type TypedArrayConstructor = {
+  new (size: number): TypedArray;
+};